requirements.txt
cachetools==5.3.2
GitPython==3.1.
packaging==23.2
python-dotenv==1.0.0
pipdeptree==2.13.0
pytest==7.4.0
tree == 0.2.4
langchain == 0.0.330
openai==0.28.1
wandb==0.15.12



